+++ lib/api/rest/functions.php
@@ -8,8 +8,9 @@
 namespace Beans\Framework\API\Rest;
 
 require_once 'breadcrumbs.php';
-require_once 'layouts.php';
+require_once 'container.php';
 require_once 'images.php';
+require_once 'layouts.php';
 require_once 'onboarding.php';
 require_once 'remove-actions.php';